Move all top-level constants and methods to Fog::AzureRM
This gem was defining some constants and methods at the top level
instead of in its own namespace.

For the constants, that's mostly just untidy: there's a small risk of
conflicts, but there's no major harm.

The methods are a different story. The top level in Ruby is an instance
of Object (because everything is an object), so defining a method here
defined it as an instance method on Object. This means that, for
instance, `get_blob_endpoint` was available as an instance method on
Integer, which is clearly nonsense.

For an extreme example in irb:

    irb(main):001:0> def send(*); 'foo'; end
    => :send
    irb(main):002:0> 1.to_s
    => "1"
    irb(main):003:0> 1.send(:to_s)
    => "foo"